+++ plugins/http/guest-js/index.ts
@@ -229,40 +229,44 @@ export async function fetch(
     rid
   })
 
-  const readableStreamBody = new ReadableStream({
-    start: (controller) => {
-      const streamChannel = new Channel<ArrayBuffer | number[]>()
-      streamChannel.onmessage = (res: ArrayBuffer | number[]) => {
-        // close early if aborted
-        if (signal?.aborted) {
-          controller.error(ERROR_REQUEST_CANCELLED)
-          return
+  // no body for 101, 103, 204, 205 and 304
+  // see https://fetch.spec.whatwg.org/#null-body-status
+  const body = [101, 103, 204, 205, 304].includes(status)
+    ? null
+    : new ReadableStream({
+        start: (controller) => {
+          const streamChannel = new Channel<ArrayBuffer | number[]>()
+          streamChannel.onmessage = (res: ArrayBuffer | number[]) => {
+            // close early if aborted
+            if (signal?.aborted) {
+              controller.error(ERROR_REQUEST_CANCELLED)
+              return
+            }
+
+            const resUint8 = new Uint8Array(res)
+            const lastByte = resUint8[resUint8.byteLength - 1]
+            const actualRes = resUint8.slice(0, resUint8.byteLength - 1)
+
+            // close when the signal to close (last byte is 1) is sent from the IPC.
+            if (lastByte == 1) {
+              controller.close()
+              return
+            }
+
+            controller.enqueue(actualRes)
+          }
+
+          // run a non-blocking body stream fetch
+          invoke('plugin:http|fetch_read_body', {
+            rid: responseRid,
+            streamChannel
+          }).catch((e) => {
+            controller.error(e)
+          })
         }
-
-        const resUint8 = new Uint8Array(res)
-        const lastByte = resUint8[resUint8.byteLength - 1]
-        const actualRes = resUint8.slice(0, resUint8.byteLength - 1)
-
-        // close when the signal to close (last byte is 1) is sent from the IPC.
-        if (lastByte == 1) {
-          controller.close()
-          return
-        }
-
-        controller.enqueue(actualRes)
-      }
-
-      // run a non-blocking body stream fetch
-      invoke('plugin:http|fetch_read_body', {
-        rid: responseRid,
-        streamChannel
-      }).catch((e) => {
-        controller.error(e)
       })
-    }
-  })
 
-  const res = new Response(readableStreamBody, {
+  const res = new Response(body, {
     status,
     statusText
   })